+++ db_criacao/04 - bling_pedidos_vendas/drawSQL-pgsql-export-2023-12-043.sql
@@ -68,8 +68,8 @@ CREATE TABLE "vendas"(
     "transporte_quantidade_volumes" INTEGER,
     "transporte_peso_bruto"         INTEGER,
     "transporte_prazo_entrega"      INTEGER,
-    "transporte_id_contato"         INTEGER             REFERENCES "contatos"("id_bling"),
-    "transporte_id_etiqueta"        BIGINT              REFERENCES "transporte_etiqueta"("id"),
+    "transporte_id_contato"         BIGINT              REFERENCES "contatos"("id_bling"),
+    "transporte_id_etiqueta"        INTEGER             REFERENCES "transporte_etiqueta"("id"),
     "alterado_em"                   TIMESTAMPTZ         DEFAULT current_timestamp
 );
 COMMENT ON COLUMN
